Refactor matching and chat UI logic, update navigation
Removed unused MatchingIngVM and refactored MatchingIng to use Scaffold and add a back button. Improved chatingRoom UI with chat bubbles, top bar, and message input. Updated chat sending logic to clear input only on success. Changed groupNavigation tab routing to include groupId. Enhanced fast match flow to navigate on success and added match state. Renamed members.kt to member.kt and added navigation route for member screen. Updated ShopModel goodId type from Int to Long.

+++ app/src/main/java/com/example/studify/data/model/ShopModel.kt
@@ -10,7 +10,7 @@ data class ShopModel (
     @SerializedName("result") @Expose var result: List<shopResult>? =null
 )  {
     data class shopResult(
-        @SerializedName("GOOD_ID") @Expose var goodId: Int? = null,
+        @SerializedName("GOOD_ID") @Expose var goodId: Long? = null,
         @SerializedName("PRICE") @Expose var price: Int? = null,
         @SerializedName("GOOD_NAME") @Expose var goodName: String? = null,
         @SerializedName("DETAIL") @Expose var detail: String? = null
